Added action bar support. Currently action buttons are not appearing. Switched to extend AppCompatActivity.
